Refactor banwebScrape into a more modular style
This modular style will be much more useful as this begins to expand
more in the near future.
